feat: extract Vault memory kernel (0.61.0)
