+++ packages/vault/src/index.mjs
@@ -1,2 +1,8 @@
 export * from './project-vault.mjs';
 export * from './vault-path-safety.mjs';
+export * from './memory-schema.mjs';
+export * from './memory-mode.mjs';
+export * from './memory-handoff.mjs';
+export * from './memory-store.mjs';
+export * from './validate-core.mjs';
+export * from './validate-memory.mjs';
